[FIX] Moved Components in folder
