+++ skill-db/frontend/src/components/DetailComponents/DetailComponent.vue
