+++ skill-db/frontend/src/components/EditComponents/EditComponent.vue
