+++ skill-db/frontend/src/views/EditView.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script lang="ts">
-import EditComponent from "@/components/EditComponent.vue";
+import EditComponent from "@/components/EditComponents/EditComponent.vue";
 
 export default {
   name: "EditView",
