+++ skill-db/frontend/src/views/NewView.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script lang="ts" >
-import EditComponent from "@/components/EditComponent.vue";
+import EditComponent from "@/components/EditComponents/EditComponent.vue";
 import {useDetailStore} from "@/stores/DetailStore";
 
 export default {
