+++ skill-db/frontend/src/views/RoleOverView.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script lang="ts">
-import RoleComponent from "@/components/RoleComponent.vue";
+import RoleComponent from "@/components/RoleComponents/RoleComponent.vue";
 
 export default {
   name: "RoleOverView",
